New Project

+++ back/src/main/java/com/unit/UnitRestController.java
@@ -1,18 +1,24 @@
 package com.unit;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 import com.GeneralRestController;
 import com.card.Card;
+import com.image.Image;
+import com.itinerary.module.Module;
 import com.relation.Relation;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.json.MappingJacksonValue;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 @RestController
 @RequestMapping("/api/units")
-public class UnitRestController extends GeneralRestController {
+public class UnitRestController extends GeneralRestController implements UnitController {
 
     @GetMapping(value = "/")
     public ResponseEntity<List<Unit>> getUnits() {
@@ -22,11 +28,7 @@ public class UnitRestController extends GeneralRestController {
     @GetMapping(value = "/{id}")
     public ResponseEntity<Unit> getUnit(@PathVariable int id) {
         Optional<Unit> unit = this.unitService.findOne(id);
-        if (unit.isPresent()) {
-            return new ResponseEntity<>(unit.get(), HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+        return unit.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PostMapping(value = "/")
@@ -64,7 +66,6 @@ public class UnitRestController extends GeneralRestController {
                 } catch (Exception e) {
                      return new ResponseEntity<>(HttpStatus.CONFLICT);
                 }
-                this.unitService.save(savedUnit.get());
                 savedUnits.add(savedUnit.get());
             } else {
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -77,10 +78,16 @@ public class UnitRestController extends GeneralRestController {
         if (!unitService.isValidName(unit)) {
             throw new Exception("Invalid name");
         }
+        if(!unit.getName().equals(savedUnit.getName())){
+            this.slideService.updateAllSlidesUnitName(savedUnit.getName(), unit.getName());
+        }
         savedUnit.setName(unit.getName());
+        savedUnit.getIncomingRelations().removeAll(savedUnit.getIncomingRelations().stream().filter(relation -> !unit.getIncomingRelations().contains(relation)).collect(Collectors.toList()));
+        savedUnit.getOutgoingRelations().removeAll(savedUnit.getOutgoingRelations().stream().filter(relation -> !unit.getOutgoingRelations().contains(relation)).collect(Collectors.toList()));
         for (Relation relation : unit.getIncomingRelations()) {
             if ((relation.getIncoming() != 0) && (relation.getOutgoing() != 0)) {
                 if (relation.getId() > 0) {
+                    relationService.save(relation);
                     savedUnit.addIncomingRelation(relation);
                 } else {
                     Relation r;
@@ -94,6 +101,7 @@ public class UnitRestController extends GeneralRestController {
         for (Relation relation : unit.getOutgoingRelations()) {
             if ((relation.getIncoming() != 0) && (relation.getOutgoing() != 0)) {
                 if (relation.getId() > 0) {
+                    relationService.save(relation);
                     savedUnit.addOutgoingRelation(relation);
                 } else {
                     Relation r;
@@ -104,31 +112,27 @@ public class UnitRestController extends GeneralRestController {
                 }
             }
         }
+        this.unitService.save(savedUnit);
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Unit> deleteUnit(@PathVariable long id) {
         Optional<Unit> unit = unitService.findOne(id);
         if (unit.isPresent()) {
-            for (Relation relation : unit.get().getOutgoingRelations()) {
-                relationService.delete(relation.getId());
+            if (unitService.ableToDeleteUnit(unit.get())) {
+                for (Relation relation : unit.get().getOutgoingRelations()) {
+                    unitService.findOne(relation.getIncoming()).map(value -> value.getIncomingRelations().remove(relation));
+                    relationService.delete(relation.getId());
+                }
+                for (Relation relation : unit.get().getIncomingRelations()) {
+                    unitService.findOne(relation.getOutgoing()).map(value -> value.getOutgoingRelations().remove(relation));
+                    relationService.delete(relation.getId());
+                }
+                unitService.delete(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.CONFLICT);
             }
-            for (Relation relation : unit.get().getIncomingRelations()) {
-                relationService.delete(relation.getId());
-            }
-            unitService.delete(id);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-
-    @DeleteMapping(value = "/relations/{id}")
-    public ResponseEntity<Relation> deleteRelation(@PathVariable long id) {
-        Optional<Relation> relation = relationService.findOne(id);
-        if (relation.isPresent()) {
-            relationService.delete(id);
-            return new ResponseEntity<>(HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -138,6 +142,12 @@ public class UnitRestController extends GeneralRestController {
     public ResponseEntity<List<Unit>> searchUnits(@RequestParam String name) {
         List<Unit> units = this.unitService.findByNameContaining(name);
         return new ResponseEntity<>(units, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/{id}/unambiguousName")
+    public ResponseEntity<Unit> getUnitUnambiguousName(@PathVariable int id) {
+        Optional<Unit> unit = this.unitService.findOne(id);
+        return unit.map(value -> new ResponseEntity<>(new Unit(unitService.getUnambiguousName(value)), HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @GetMapping(value = "/{id}/absoluteName")
@@ -165,6 +175,77 @@ public class UnitRestController extends GeneralRestController {
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+    }
+
+    @GetMapping(value = "/{id}/images")
+    public ResponseEntity<MappingJacksonValue> getUnitImages(@PathVariable int id) {
+        Optional<Unit> optionalUnit = this.unitService.findOne(id);
+        if (optionalUnit.isPresent()) {
+            MappingJacksonValue result = new MappingJacksonValue(optionalUnit.get().getImages());
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping(value="/{unitId}/images/{imageId}")
+    public ResponseEntity<Image> getImage(@PathVariable int unitId, @PathVariable int imageId) {
+        Optional<Unit> optionalUnit = this.unitService.findOne(unitId);
+        if (optionalUnit.isPresent()) {
+            Optional<Image> image = this.imageService.findOne(imageId);
+            return (image.isPresent() && image.get().getUnitId() == unitId)?(new ResponseEntity<>(image.get(), HttpStatus.OK)):(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+    }
+
+    @PostMapping(value = "/{unitId}/images", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<Image> addImage(@RequestParam(value = "imageFile") MultipartFile imageFile, @PathVariable int unitId) {
+        Optional<Unit> optionalUnit = this.unitService.findOne(unitId);
+        if (optionalUnit.isPresent()) {
+            Image newImage = new Image(imageFile.getOriginalFilename());
+            newImage.setUnitId(unitId);
+            this.imageService.save(newImage);
+            this.imageService.setImage(newImage, imageFile);
+            this.imageService.save(newImage);
+
+            optionalUnit.get().addImage(newImage);
+            this.unitService.save(optionalUnit.get());
+
+            return new ResponseEntity<>(newImage, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @DeleteMapping(value="/{unitId}/images/{imageId}")
+    public ResponseEntity<Image> deleteImage(@PathVariable int unitId, @PathVariable int imageId) {
+        Optional<Unit> optionalUnit = this.unitService.findOne(unitId);
+        if (optionalUnit.isPresent()) {
+            Optional<Image> image = this.imageService.findOne(imageId);
+            if(image.isPresent()) {
+                this.imageService.delete(imageId);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+    }
+
+    @GetMapping(value="/module/{id}")
+    public ResponseEntity<Unit> getModuleUnit(@PathVariable long id){
+        Optional<Module> module = this.moduleService.findOne(id);
+        if (module.isPresent()) {
+            Optional<Unit> result = this.unitService.findOne(this.unitService.findModuleUnit(id));
+            if(result.isPresent()){
+                return new ResponseEntity<>(result.get(), HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
 }
